refactor and add script
